Default to printing progress bar to sterr, rather than stdout.
As per https://github.com/cloud-oak/ProgressBars.jl/issues/46, the
correct default behavior of this package should be to print to
stderr. This commit brings us in-line with other ProgressBar-style
utilities, like `pv` and `tqdm`. It also allows ProgressBar.jl to
be dropped into piped scripts.

By default, ProgressBar will print to stderr. In a normal terminal,
ProgressBar will look the same as it used to. To set ProgressBar to
print to stdout and behave as it used to, ProgressBar now accepts a
`output_stream` argument, which can be set to `stdout`. E.g.:
`ProgressBar(1:5, output_stream=stdout)`